everything is single page, fixed footer
