+++ components/footer.js
@@ -3,9 +3,7 @@ import React from 'react';
 export default function Footer() {
     const footerstyle = {
         position: "absolute",
-        left:0,
-        bottom:0,
-        right:0,
+        width: "100%"
     }
     return (
         <footer id="footer" style={footerstyle}>
