+++ styles/globals.css
@@ -2,10 +2,10 @@ html,
 body {
   padding: 0;
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+  font-family: 'Courier Prime', monospace; 
   color: #101820;
   background-color: #f0f0dc;
+
 }
 
 a {
@@ -28,7 +28,6 @@ a {
 }
 
 #emily {
-  font-family: 'Courier Prime', monospace;
   font-size: 300%
 }
 
@@ -38,7 +37,6 @@ a {
 
 #blurb, #intro {
   width: 80%;
-  font-family: 'Courier Prime', monospace;
   font-size: 140%;
 }
 
@@ -84,9 +82,7 @@ h1 {
   margin: 70px 200px;
 }
 
-#choose {
-  font-family: 'Press Start 2P';
-}
+
 
 #fighter {
   font-family: 'Courier Prime', monospace;
